requirements.txt
langchain==0.3.12
langchain-community==0.3.12
langchain-core==0.3.25
langchain-cohere==0.3.2
langchain-qdrant==0.2.0
cohere==5.11.4
qdrant-client==1.12.1
streamlit==1.40.2
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
tenacity==9.0.0
typing-extensions==4.12.2
pydantic==2.9.2
pypdf==3.17.1